refactor: update wallet-aware-routes to app-routes
