+++ frontend/src/App.tsx
@@ -12,7 +12,7 @@ import { API_URL } from '@/config';
 import { solanaNetworkNameFromCluster } from '@constants/chains';
 import { DynamicProvider } from '@contexts/dynamic';
 import { petsContractParams } from '@/petsContractParams';
-import { WalletAwareRoutes } from '@router';
+import { AppRoutes } from '@router';
 import './App.css';
 
 const solanaNetwork = solanaNetworkNameFromCluster(import.meta.env.VITE_SOLANA_CLUSTER);
@@ -29,7 +29,7 @@ const App: React.FC = () => {
                                 <AuthProvider>
                                     <PetsConfigProvider evm={petsContractParams}>
                                         <BrowserRouter>
-                                            <WalletAwareRoutes />
+                                            <AppRoutes />
                                         </BrowserRouter>
                                     </PetsConfigProvider>
                                 </AuthProvider>
