+++ frontend/src/router/index.ts
@@ -1,6 +1,5 @@
-export { useAppLoggedIn } from '@router/useAppLoggedIn';
-export { default as BattleRoute } from '@router/battle-route';
-export { default as BreedRoute } from '@router/breed-route';
-export { default as LevelUpRoute } from '@router/level-up-route';
-export { default as RenameRoute } from '@router/rename-route';
-export { default as WalletAwareRoutes } from '@router/wallet-aware-routes';
+export { default as AppRoutes } from '@router/app-routes';
+export { default as BattleRoute } from '@router/routes/battle-route';
+export { default as BreedRoute } from '@router/routes/breed-route';
+export { default as LevelUpRoute } from '@router/routes/level-up-route';
+export { default as RenameRoute } from '@router/routes/rename-route';
