+++ frontend/src/router/routes/battle-route/index.tsx
